Time: 43 ms (29.16%), Space: 61.8 MB (6.45%) - LeetHub

+++ 0518-coin-change-ii/0518-coin-change-ii.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int change(int amount, vector<int>& coins) {
         int n = coins.size();
-        vector<vector<int>> dp(n+1, vector<int>(amount+1, 0));
+        vector<vector<long long>> dp(n+1, vector<long long>(amount+1, 0));
         for(int i = 0; i <= n; i++) dp[i][0] = 1;
         for(int i = 1; i <= n; i++){
             for(int j = 0; j <= amount; j++){
@@ -12,7 +12,7 @@ public:
                     take = dp[i][j - coins[i-1]];
                 }
 
-                dp[i][j] = skip + take;
+                dp[i][j] = (long long)skip + (long long)take;
             }
         }
         return dp[n][amount];
